backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.12
anthropic==0.40.0
python-dotenv==1.0.1
pydantic==2.10.3
aiofiles==24.1.0
pdfplumber==0.11.4
pillow==11.0.0
requests==2.32.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
